xwayland: simplify unmanaged views
They are no longer toplevel views, so they don't need most of the
operations.
